Green Light |

+++ softwareprocess/dispatch.py
@@ -263,7 +263,7 @@ def degreeToFloat(degree):
     return degree
 
 def degreeToString(degree):
-    minute = str("{:.1f}".format((degree - float(degree)) * 60))
+    minute = str("{:.1f}".format((degree - int(degree)) * 60))
     if '-' in minute:
         minute = minute.replace('-', '')
     minute = minute.split('.')
